Block Comments: Extract CRUD actions into a separate hook
